docs: Upload Both Reports in the Correct File Extension
In a previous commit, I added the reports in .ipynb. However,
Udacity's project rubric requires that they must be either .html
or .pdf to make the report more accessible.

I also removed the instruction text from the both of them as a
suggestion from the first reviewer, as to make them look more polished.

In the case of the act_report, I also added some screenshots from
selected tweets to make it more engaging to my readers.